NMS-9128: working filter tests
